requirements-dev.txt
ruff==0.1.13
black==23.12.1

mysql-connector-python==8.2.0
